Move get schema down
