+++ snuba/clickhouse/http.py
@@ -55,7 +55,7 @@ class HTTPBatchWriter(BatchWriter):
         self.__pool = HTTPConnectionPool(host, port)
         self.__options = options if options is not None else {}
         self.__table_name = table_name or schema \
-            .get_writable_storage() \
+            .get_storage() \
             .get_table_name()
         self.__chunk_size = chunk_size
         self.__encoder = encoder
