+++ tests/base.py
@@ -189,7 +189,7 @@ class BaseEventsTest(BaseDatasetTest):
         super(BaseEventsTest, self).setup_method(test_method, 'events')
         self.table = self.dataset.get_dataset_schemas() \
             .get_write_schema() \
-            .get_writable_storage() \
+            .get_storage() \
             .get_table_name()
         self.event = get_event()
