+++ tests/test_perf.py
@@ -10,7 +10,7 @@ class TestPerf(BaseEventsTest):
         table = dataset \
             .get_dataset_schemas() \
             .get_write_schema() \
-            .get_writable_storage() \
+            .get_storage() \
             .get_local_table_name()
 
         assert self.clickhouse.execute("SELECT COUNT() FROM %s" % table)[0][0] == 0
